% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{kNN classifier}
\usage{
classify(distMatrix, k, dset_train, r1, cc)
}
\arguments{
\item{distMatrix}{The distance matrix you are using to caluclate distances between observations.}

\item{k}{Number of k nearest neighbors you want to identify for classifying}

\item{dset_train}{The name of the data frame or matrix with which you are training your classifier}

\item{r1}{First row from which you want to start identifying nearest neighbors; 1 if training, 0 if scoring}

\item{cc}{The column in your training dataset that has the classifications of each observation}
}
\value{
predictedClasses Return a vector of classified values for each observation in the prediction dataset
}
\description{
Classify the input with a k nearest neighbors classifier.
}
\examples{
# create artificial dataset
inputsTest   <- matrix(rnorm(100), ncol=2)
inputsTrain  <- matrix(rnorm(100), ncol=2)
classesTrain <- cbind(inputsTrain, c(rep(0, 25), rep(1, 25)))
c1 <- 1
c2 <- 2
p <- 2
distance_matrix<- distance(classesTrain, c1, c2, classesTrain, p)
# get the kNN predictions for the test set
names(classesTrain)<-c(V1="weight", V2="height", Animal="Cat")
k <- 1
r1 <- 1
cc <- 3
classes<- classify(distance_matrix, k, classesTrain, r1, cc)
}

