% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Calculates distances for KNN}
\usage{
distance(dset_pred, c1, c2, dset_train, p)
}
\arguments{
\item{dset_pred}{A data frame or a matrix where rows are observations and
columns are features. If you are training a dataset this is your training dataset,
if you are predicting a validation file, this is your validation file.}

\item{c1}{Beginning column number for features in your dset_pred (assumes all features are in adjacent columns)}

\item{c2}{Ending column number for features in your dset_pred   (assumes all features are in adjacent columns)}

\item{dset_train}{The name of the data frame or matrix with which you are training your classifier}

\item{p}{Distance metric the classifier should use, the value can be
either 1(L1 norm), 2(euclidian) or Inf.}
}
\value{
distMatrix A matrix of computed distances.
}
\description{
Classify distances bewteen observations in the training data or in the training vs validation.
}
\examples{
# create artificial dataset
inputsTest   <- matrix(rnorm(100), ncol=2)
inputsTrain  <- matrix(rnorm(100), ncol=2)
classesTrain <- cbind(inputsTrain, c(rep(0, 25), rep(1, 25)))
# get the kNN predictions for the test set
names(classesTrain)<-c(V1="weight", V2="height", Animal="Cat")
test<-distance(classesTrain, 1, 2, classesTrain, 2)
}

